feat: trade page pool header mobile optimization
